Properly detect the license of certain Microsoft packages.
