+++ Src/PackageGuard.Core/CSharpProjectAnalyzer.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using NuGet.ProjectModel;
+using PackageGuard.Core.Common;
 using Pathy;
 
 namespace PackageGuard.Core;
