+++ Src/PackageGuard.Core/CSharpProjectScanner.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Build.Construction;
 using Microsoft.Extensions.Logging;
+using PackageGuard.Core.Common;
 using Pathy;
 
 namespace PackageGuard.Core;
