+++ Src/PackageGuard.Core/FetchingStrategies/IFetchLicense.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+
+namespace PackageGuard.Core;
+
+/// <summary>
+/// Represents a strategy for fetching a license for a package.
+/// </summary>
+public interface IFetchLicense
+{
+    Task FetchLicenseAsync(PackageInfo package);
+}
