+++ Src/PackageGuard.Specs/PackageGuard.Specs.csproj
@@ -16,7 +16,6 @@
         <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       </PackageReference>
       <PackageReference Include="Meziantou.Extensions.Logging.InMemory" Version="1.2.6" />
-      <PackageReference Include="Meziantou.Extensions.Logging.Xunit" Version="1.0.13" />
       <PackageReference Include="Microsoft.Extensions.Diagnostics.Testing" Version="9.6.0" />
       <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.14.1" />
       <PackageReference Include="MSTest" Version="3.9.3" />
@@ -24,7 +23,6 @@
         <PrivateAssets>all</PrivateAssets>
         <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       </PackageReference>
-      <PackageReference Include="xunit.extensibility.core" Version="2.9.3" />
       </ItemGroup>
 
     <ItemGroup>
